code/photo_researcher/requirements.txt
openai
psycopg2-binary
python-dotenv
requests
beautifulsoup4
